--- src/pages/forgot-password/forgot-password.module.css
@@ -1,15 +1,0 @@
-.wrapper{
-  display: flex;
-  align-items: center;
-  justify-content: center; 
-  height: 100vh;
-}
-.form{
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-}
-
-.form > div, button{
-  margin-top: calc(var(--offset-base-size)  * 4);
-}
