+++ src/pages/index.tsx
@@ -1,9 +1,9 @@
-import {ForgotPasswordPage} from './forgot-password/forgot-password';
+import {ForgotPasswordPage} from './forgot-password';
 import {LoginPage} from './login';
 import { MainPage } from './main/main-page';
 import { ProfilePage } from './profile/profile';
-import { RegistrationPage } from './registation/registration';
-import { ResetPasswordPage } from './reset-password/reset-password';
+import { RegistrationPage } from './registration';
+import { ResetPasswordPage } from './reset-password';
 import { Order } from './order';
 import { PageNotFound } from './page-not-found/page-not-found';
