+++ src/pages/registration.tsx
@@ -1,4 +1,3 @@
-import styles from './registration.module.css';
 import {
   Input,
   Button,
@@ -6,19 +5,15 @@ import {
 import { useState, useEffect } from 'react';
 import { Link, Redirect } from 'react-router-dom';
 import { useDispatch, useSelector } from 'react-redux';
-import { registrationUser } from '../../services/auth/actions';
-import { TUser } from '../../types/types';
+import { registrationUser } from '../services/auth/actions';
+import { Form } from '../components/form/form';
+import { TUser } from '../types/types';
 
 export const RegistrationPage: React.FC = () => {
   const dispatch = useDispatch();
   const user = useSelector((store: any) => store.auth.user);
   const [value, setValue] = useState<TUser>({ name: '', email: '', password: '' });
   const [isPasswordShow, setIsPasswordShow] = useState<boolean>(false);
-
-  useEffect(() => {
-    const button = document.getElementById('enterButton');
-    button?.children[0].setAttribute('type', 'submit');
-  }, []);
 
   const onChange = (e: React.ChangeEvent<HTMLInputElement>) => {
     setValue({ ...value, [e.target.name]: e.target.value });
@@ -38,8 +33,7 @@ export const RegistrationPage: React.FC = () => {
   }
 
   return (
-    <div className={styles.wrapper}>
-      <form className={styles.form}>
+      <Form onSubmit={onSubmit} idButton='enterButton'>
         <h1 className="text text_type_main-medium">Регистрация</h1>
         <Input
           type="text"
@@ -78,7 +72,6 @@ export const RegistrationPage: React.FC = () => {
             &nbsp;Войти
           </Link>
         </span>
-      </form>
-    </div>
+      </Form>
   );
 };
